// lib/api/model/createModuleCommand.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CreateModuleCommand {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  courseOfferingId: number | string;
  title: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  orderIndex: number | string;
}